reorganize test: code gen
